fix a couple of warnings on win64

+++ vm/code_heap.cpp
@@ -5,7 +5,7 @@ namespace factor
 
 code_heap::code_heap(cell size)
 {
-	if(size > (1L << (sizeof(cell) * 8 - 6))) fatal_error("Heap too large",size);
+	if(size > ((u64)1 << (sizeof(cell) * 8 - 6))) fatal_error("Heap too large",size);
 	seg = new segment(align_page(size),true);
 	if(!seg) fatal_error("Out of memory in heap allocator",size);
 	allocator = new free_list_allocator<code_block>(size,seg->start);
